Move errors to exceptions and cleanup
