+++ lib/jsonapi_plug/exceptions.ex
@@ -1,12 +1,43 @@
 defmodule JSONAPIPlug.Exceptions do
   @moduledoc false
 
+  defmodule InvalidDocument do
+    @moduledoc """
+    Defines a generic exception for when an invalid document is received.
+    """
+    defexception message: nil, reference: nil
+
+    @spec exception(keyword()) :: Exception.t()
+    def exception(options) do
+      message = Keyword.fetch!(options, :message)
+      reference = Keyword.fetch!(options, :reference)
+
+      %__MODULE__{message: message, reference: reference}
+    end
+  end
+
+  defmodule InvalidHeader do
+    @moduledoc """
+    Defines a generic exception for when an invalid header is received.
+    """
+    defexception header: nil, message: nil, reference: nil, status: nil
+
+    @spec exception(keyword()) :: Exception.t()
+    def exception(options) do
+      header = Keyword.fetch!(options, :header)
+      message = Keyword.fetch!(options, :message)
+      reference = Keyword.fetch!(options, :reference)
+      status = Keyword.fetch!(options, :status)
+
+      %__MODULE__{header: header, message: message, reference: reference, status: status}
+    end
+  end
+
   defmodule InvalidQuery do
     @moduledoc """
     Defines a generic exception for when an invalid query parameter is received.
     """
-    defexception plug_status: 400,
-                 message: "invalid query",
+    defexception message: "invalid query",
                  type: nil,
                  param: nil,
                  value: nil
@@ -23,20 +54,6 @@ defmodule JSONAPIPlug.Exceptions do
         param: param,
         value: value
       }
-    end
-  end
-
-  defmodule InvalidDocument do
-    @moduledoc """
-    Defines a generic exception for when an invalid document is received.
-    """
-    defexception plug_status: 400, message: nil, reference: nil
-
-    @spec exception(keyword()) :: Exception.t()
-    def exception(options) do
-      message = Keyword.fetch!(options, :message)
-      reference = Keyword.fetch!(options, :reference)
-      %__MODULE__{message: message, reference: reference}
     end
   end
 end
