+++ lib/jsonapi_plug/plug/document.ex
@@ -6,7 +6,7 @@ defmodule JSONAPIPlug.Plug.Document do
   and stores it in the `Plug.Conn` private assigns for later use by other plugs and application code.
   """
 
-  alias JSONAPIPlug.Document
+  alias JSONAPIPlug.{Document, Normalizer}
   alias Plug.Conn
 
   @behaviour Plug
@@ -24,10 +24,13 @@ defmodule JSONAPIPlug.Plug.Document do
           conn,
         _opts
       ) do
+    document = Document.deserialize(body_params)
+    body_params = Normalizer.denormalize(document, jsonapi_plug.view, conn)
+
     Conn.put_private(
       conn,
       :jsonapi_plug,
-      %JSONAPIPlug{jsonapi_plug | document: Document.deserialize(body_params)}
+      %JSONAPIPlug{jsonapi_plug | document: document, params: body_params}
     )
   end
 end
