--- lib/jsonapi_plug/plug/params.ex
@@ -1,33 +1,0 @@
-defmodule JSONAPIPlug.Plug.Params do
-  @moduledoc """
-  Transforms conn body params to denormalized form
-
-  This plug uses the configured normalizer (which defaults to `JSONAPIPlug.Normalizer.Ecto`) to
-  transform the received `JSON:API` document into the format suitable for application use.
-
-  The returned data replaces `Plug.Conn` `body_params` and is added to `params` under a "data" key.
-  """
-
-  alias JSONAPIPlug.API
-  alias Plug.Conn
-
-  @behaviour Plug
-
-  @impl Plug
-  def init(options), do: options
-
-  @impl Plug
-  def call(%Conn{private: %{jsonapi_plug: %JSONAPIPlug{} = jsonapi_plug}} = conn, _options) do
-    normalizer = API.get_config(jsonapi_plug.api, [:normalizer])
-    body_params = normalizer.denormalize(jsonapi_plug.document, jsonapi_plug.view, conn)
-
-    %Conn{
-      conn
-      | body_params: body_params,
-        params:
-          conn.params
-          |> Map.drop(["errors", "included", "jsonapi", "links", "meta"])
-          |> Map.put("data", body_params)
-    }
-  end
-end
